+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Removed
 
+- Remove support for `php:^7.4`
 - Remove support for `league/flysystem:^2`
 
 ## 2.2.0 - 2022-10-28